+++ src/main/java/DecisionIQ/Backend/model/RecommendationExplainer.java
@@ -9,25 +9,34 @@ import DecisionIQ.Backend.model.UserPreference;
 
 public class RecommendationExplainer {
 
+
     public List<String> explain(
             Product product,
             UserPreference preference
     ) {
 
+
         List<String> reasons = new ArrayList<>();
+
 
         for (Map.Entry<String, Integer> entry :
                 preference.getPreferences().entrySet()) {
 
+
             String feature = entry.getKey();
+
 
             int importance = entry.getValue();
 
-            int productScore =
+
+            double productScore =
                     product.getFeatureScore(feature);
+
+
 
             if (importance >= 70 &&
                     productScore >= 85) {
+
 
                 reasons.add(
                         feature +
@@ -35,18 +44,26 @@ public class RecommendationExplainer {
                         productScore +
                         "/100)"
                 );
+
             }
+
+
 
             if (importance >= 70 &&
                     productScore < 60) {
+
 
                 reasons.add(
                         feature +
                         " is weaker compared to your requirement"
                 );
+
             }
+
         }
+
 
         return reasons;
     }
+
 }
